first attempt

+++ p2_continuous-control/Critic.py
@@ -35,6 +35,6 @@ class Critic(nn.Module):
     def forward(self, state, action):
         """Build a critic (value) network that maps (state, action) pairs -> Q-values."""
         xs = F.relu(self.fcs1(state))
-        x = torch.cat((xs, action), dim=1) # Should work with actor's output here.
+        x = torch.cat((xs, action.float()), dim=1)  # Should work with actor's output here.
         x = F.relu(self.fc2(x))
         return self.fc3(x)
